對齊人物選項，修正錯誤

+++ GameGlobal/GameGlobal/GlobalVariables.cs
@@ -766,7 +766,7 @@
             }
             try
             {
-                CreatedOfficerAbilityFactor = int.Parse(nextSibling.Attributes.GetNamedItem("CreatedOfficerAbilityFactor").Value);
+                CreatedOfficerAbilityFactor = float.Parse(nextSibling.Attributes.GetNamedItem("CreatedOfficerAbilityFactor").Value);
             }
             catch (Exception exception24)
             {
